// packages/vod/src/models/get-execution-metas.ts
export const getExecutionMetas: Record<string, ModelMeta> = {
  GetExecutionRequest: {
    attributeMap: { runId: "RunId" },
    swaggerTypes: { runId: "str" },
  },
  GetExecutionResponse: {
    attributeMap: { code: "Code", control: "Control", input: "Input", meta: "Meta", operation: "Operation", output: "Output", runId: "RunId", status: "Status" },
    swaggerTypes: { code: "str", control: "ControlForGetExecutionOutput", input: "InputForGetExecutionOutput", meta: "MetaForGetExecutionOutput", operation: "OperationForGetExecutionOutput", output: "OutputForGetExecutionOutput", runId: "str", status: "str" },
  },
  AudioStreamMetaForGetExecutionOutput: {
    attributeMap: { bitrate: "Bitrate", codec: "Codec", duration: "Duration", quality: "Quality", sampleRate: "SampleRate" },
    swaggerTypes: { bitrate: "int", codec: "str", duration: "float", quality: "str", sampleRate: "int" },
  },
  ControlForGetExecutionOutput: {
    attributeMap: { callbackArgs: "CallbackArgs", clientToken: "ClientToken", priority: "Priority" },
    swaggerTypes: { callbackArgs: "str", clientToken: "str", priority: "int" },
  },
  ConvertEnhanceForGetExecutionOutput: {
    attributeMap: { audioStreamMeta: "AudioStreamMeta", createTime: "CreateTime", duration: "Duration", dynamicRange: "DynamicRange", encodedType: "EncodedType", encrypt: "Encrypt", encryption: "Encryption", fileId: "FileId", fileType: "FileType", format: "Format", logoType: "LogoType", md5: "Md5", size: "Size", storeUri: "StoreUri", tosStorageClass: "TosStorageClass", videoStreamMeta: "VideoStreamMeta" },
    swaggerTypes: { audioStreamMeta: "AudioStreamMetaForGetExecutionOutput", createTime: "str", duration: "float", dynamicRange: "str", encodedType: "str", encrypt: "bool", encryption: "EncryptionForGetExecutionOutput", fileId: "str", fileType: "str", format: "str", logoType: "str", md5: "str", size: "float", storeUri: "str", tosStorageClass: "str", videoStreamMeta: "VideoStreamMetaForGetExecutionOutput" },
  },
  ConvertTaskForGetExecutionOutput: {
    attributeMap: { enhance: "Enhance", type: "Type" },
    swaggerTypes: { enhance: "ConvertEnhanceForGetExecutionOutput", type: "str" },
  },
  EncryptionForGetExecutionOutput: {
    attributeMap: { kid: "Kid" },
    swaggerTypes: { kid: "str" },
  },
  EnhanceForGetExecutionOutput: {
    attributeMap: { modules: "Modules", moeEnhance: "MoeEnhance", type: "Type" },
    swaggerTypes: { modules: "list[ModuleForGetExecutionOutput]", moeEnhance: "MoeEnhanceForGetExecutionOutput", type: "str" },
  },
  InputForGetExecutionOutput: {
    attributeMap: { type: "Type", vid: "Vid" },
    swaggerTypes: { type: "str", vid: "str" },
  },
  MetaForGetExecutionOutput: {
    attributeMap: { createTime: "CreateTime", endTime: "EndTime", spaceName: "SpaceName", startTime: "StartTime", trigger: "Trigger" },
    swaggerTypes: { createTime: "str", endTime: "str", spaceName: "str", startTime: "str", trigger: "str" },
  },
  ModuleForGetExecutionOutput: {
    attributeMap: { type: "Type" },
    swaggerTypes: { type: "str" },
  },
  MoeEnhanceForGetExecutionOutput: {
    attributeMap: { config: "Config", target: "Target", videoStrategy: "VideoStrategy" },
    swaggerTypes: { config: "str", target: "TargetForGetExecutionOutput", videoStrategy: "VideoStrategyForGetExecutionOutput" },
  },
  OperationForGetExecutionOutput: {
    attributeMap: { task: "Task", type: "Type" },
    swaggerTypes: { task: "TaskForGetExecutionOutput", type: "str" },
  },
  OutputForGetExecutionOutput: {
    attributeMap: { task: "Task", type: "Type" },
    swaggerTypes: { task: "ConvertTaskForGetExecutionOutput", type: "str" },
  },
  TargetForGetExecutionOutput: {
    attributeMap: { bitrate: "Bitrate", fps: "Fps", res: "Res", resLimit: "ResLimit", scaleRatio: "ScaleRatio" },
    swaggerTypes: { bitrate: "int", fps: "float", res: "str", resLimit: "int", scaleRatio: "float" },
  },
  TaskForGetExecutionOutput: {
    attributeMap: { enhance: "Enhance", type: "Type" },
    swaggerTypes: { enhance: "EnhanceForGetExecutionOutput", type: "str" },
  },
  VideoStrategyForGetExecutionOutput: {
    attributeMap: { repairStrength: "RepairStrength", repairStyle: "RepairStyle" },
    swaggerTypes: { repairStrength: "int", repairStyle: "int" },
  },
  VideoStreamMetaForGetExecutionOutput: {
    attributeMap: { bitrate: "Bitrate", codec: "Codec", definition: "Definition", duration: "Duration", fps: "Fps", height: "Height", width: "Width" },
    swaggerTypes: { bitrate: "int", codec: "str", definition: "str", duration: "float", fps: "float", height: "int", width: "int" },
  },
};